feat: add support for processing and modifying SquashFS files in launch script

+++ launch.sh
@@ -182,6 +182,7 @@ update_file_shebang() {
         echo "#!/usr/bin/env bash" > "$file.new"
         cat "$file.tmp" >> "$file.new"
         mv "$file.new" "$file"
+        chmod +x "$file"
         rm -f "$file.tmp"
     else
         echo "No need to update shebang for $file"
@@ -210,36 +211,39 @@ modify_squashfs_shebang() {
     squashfs_file="$1"
     tmpdir=$(mktemp -d) || return 1
 
-    unsquashfs -d "$tmpdir" "$squashfs_file" || {
+    echo "Modifying shebangs in $squashfs_file"
+    if ! unsquashfs -d "$tmpdir" "$squashfs_file"; then
         echo "Failed to extract squashfs"
         rm -rf "$tmpdir"
         return 1
-    }
+    fi
 
     find_shell_scripts "$tmpdir" | update_shebangs_from_list
 
-    mksquashfs "$tmpdir" "$squashfs_file" -noappend -comp xz > /dev/null 2>&1 || {
+    echo "Rebuilding squashfs file $squashfs_file"
+    rm -f "$squashfs_file"
+    if !  mksquashfs "$tmpdir" "$squashfs_file" -noappend -comp xz; then
         echo "Failed to rebuild squashfs"
         rm -rf "$tmpdir"
         return 1
-    }
+    fi
 
     rm -rf "$tmpdir"
-    echo "SquashFS modified successfully."
 }
 
 process_squashfs_files() {
     search_dir="$1"
 
+    echo "Processing SquashFS files in $search_dir"
     find "$search_dir" -type f -name "*.squashfs" | while read -r squashfs_file; do
-        completed_file="${squashfs_file}.completed"
-        if [ -f "$completed_file" ]; then
-            echo "Skipping $squashfs_file (already completed)"
+        processed_marker="${squashfs_file}.processed"
+        if [ -f "$processed_marker" ]; then
+            echo "Skipping $squashfs_file, already processed"
             continue
         fi
         echo "Processing $squashfs_file"
         if modify_squashfs_shebang "$squashfs_file"; then
-            touch "$completed_file"
+            touch "$processed_marker"
         else
             echo "Failed to process $squashfs_file"
         fi
@@ -358,7 +362,7 @@ main() {
     python3 "$PAK_DIR/src/replace_string_in_file.py" "$EMU_DIR/control.txt" \
         TEMP_DATA_DIR "${TEMP_DATA_DIR#/}"
 
-    process_squashfs_files "$EMU_DIR/packs"
+    process_squashfs_files "$EMU_DIR/libs"
 
     minui-power-control &
